+++ doc/whats-new.md
@@ -2,8 +2,36 @@ What's new
 ==========
 
 
-0.1.1
------
+0.1.4 (21st June 2020)
+----------------------
+
+### Internal changes
+
+- Make compatible with v5.13 of pyside2 (#45)\
+  By [John Omotani](https://github.com/johnomotani)
+
+
+0.1.3 (20th June 2020)
+----------------------
+
+### Internal changes
+
+- Warn instead of failing in case of ImportError when setting "Qt5Agg" backend
+  (#44)\
+  By [John Omotani](https://github.com/johnomotani)
+
+
+0.1.2 (20th June 2020)
+----------------------
+
+### Internal changes
+
+- Use argparse in command line scripts (#42)\
+  By [John Omotani](https://github.com/johnomotani)
+
+
+0.1.1 (9th June 2020)
+---------------------
 
 ### New features